requirements.txt
psycopg2-binary
pandas
openpyxl
python-dotenv
boto3